Move SolverL1RLS out of the examples directory, and convert TFOCSSuite to SolverL1RLSSuite, calling SolverL1RLS directly.
